+++ src/Mapbender/ManagerBundle/Resources/public/js/application-edit.js
@@ -83,12 +83,14 @@ $(function() {
         items: "tr.sourceinst[data-href]",
         distance: 20,
         stop: function(event, ui) {
+            var $item = $(ui.item);
+            var $siblings = $('tr:not(.dummy)', $item.parent());
             $.ajax({
-                url: $(ui.item).attr("data-href"),
+                url: $item.attr("data-href"),
                 type: "POST",
                 data: {
-                    number: ui.item.index(),
-                    new_layersetId: ui.item.closest('table.layersetTable[data-id]').attr("data-id")
+                    number: $siblings.index($item),
+                    new_layersetId: $item.closest('table.layersetTable[data-id]').attr("data-id")
                 }
             });
         }